+++ src/flag_gems/ops/__init__.py
@@ -67,8 +67,7 @@ from .groupnorm import group_norm
 from .gt import gt, gt_scalar
 from .hstack import hstack
 from .index_add import index_add
-from .index_put import index_put
-from .index_put_ import index_put_
+from .index_put import index_put, index_put_
 from .index_select import index_select
 from .instancenorm import instance_norm
 from .isclose import allclose, isclose
